+++ src/index.ts
@@ -125,7 +125,7 @@ export type { Props as DeviceSelectorProps } from './Device/DeviceSelector/Devic
 
 export type { PaneProps } from './App/App';
 export type { DfuImage } from './Device/initPacket';
-export type { DeviceSetup, IDeviceSetup } from './Device/deviceSetup';
+export type { DeviceSetupConfig, DeviceSetup } from './Device/deviceSetup';
 export { prepareDevice } from './Device/deviceSetup';
 export type {
     AppInfo,